requirements.txt
langgraph
langchain-core
langchain-google-genai
pandas
flask
python-dotenv
